requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
pandas
requests
apache-airflow-providers-http